Move utility functions; add scaffolding
